Simplify requirements and make the transformer convert only list of user
In its original version the requirements were kind of confusing. It seemed
there were two tasks. One was to convert from e.g. JSON to XML and vice
versa. The other was to change the format or representation of the user
by parsing the name field and splitting it into a nested object with
separate fields for given name and family name. It also had

Furthermore the converter had to convert both a single object in an unnamed
array, whereas the other use case was to convert a list of user objects
nested under a "users" object. This just complicated things in this MVP.

Now the requirements are simpler and hence the implementation can be
simpler in its first version. Also, the spec can now describe in one
Scenario outline the four combinations of input and output formats with
one example table.

The original requirement to order the elements by "sequence" in the
exported file is retained.